requirements.txt
Flask-WTF
Flask-SQLAlchemy
psycopg2
Flask-Migrate
